docker/vulnerable-app/requirements.txt
Flask==2.3.0
PyYAML==6.0
Werkzeug==2.3.0
